# PERF: Optimize Reader.TypeDeserializerCache.cs