@bdlane 2022-02-19
Hi @perf-cat 

Thanks for the contribution.

I see you've made a number of similar contributions to several other repos in the past few days, all of which are small edits around performance, and are measured by BenchmarkDotNET. Just wondering what the deal is with this?